pharamacist section completed

+++ hospitalmanagement.py
@@ -1,6 +1,6 @@
 import functools
-from flask import Flask, render_template,redirect, request, Blueprint, flash,session, url_for, abort, jsonify, json
-from .dbschema import UserStore, PatientStore, PatientMed
+from flask import Flask, render_template,redirect, request, Blueprint, flash,session, url_for, abort, jsonify
+from .dbschema import UserStore, PatientStore, PatientMed, MedicineMaster
 from datetime import datetime
 from werkzeug.security import generate_password_hash
 
@@ -87,21 +87,50 @@ def pat_details():
 @login_required
 def pharm_pat_details():
 
+    pat = None
+    meds= None
     if request.method=='GET':
 
         id = request.args.get('id')
-        pat = PatientStore.objects(ws_pat_id=id)
-        meds = PatientMed.objects(pat_id= pat)
+
+        if id != None:
+            pat = PatientStore.objects(ws_pat_id=id).first()
+            meds = PatientMed.objects(pat_id= pat)
 
         if pat== None and id is not None:
-            flash('No Patient Found')
-        elif(id==None):
-            return(400)
+            flash('No Patient Found',"danger")
+        
 
-        return render_template('pharma.html', meds= meds)
+        return render_template('pharma.html', pat= pat, meds= meds)
 
+
+@bp.route('/pharmacy/issuemeds', methods=['GET','POST'])
+@login_required
+def issueMedicine():
     
+    if request.method== 'GET':
+        return render_template('issue_med.html', pat= {'pat_id':request.args.get('issue_medicine')})
 
+    if request.method== 'POST':
+
+        req =request.get_json(force=True)
+        if 'med_avail' in req.keys():
+
+            med= MedicineMaster.objects(med_name= req["med_name"]).first()
+            return jsonify(med)
+
+        elif 'submit_button' in req.keys():
+            pat_id = req["pat_id"]
+            pat = PatientStore.objects(ws_pat_id= pat_id).first()
+
+            for meds in req["data"]:
+                mm = MedicineMaster.objects(med_id= meds['med_id']).first()
+                PatientMed(pat_id= pat, med_id = mm, med_qty_issued= meds['qty_issued']).save()
+                MedicineMaster.objects(med_id= meds['med_id']).update_one(dec__med_qty= meds['qty_issued'])
+
+            return url_for('hospitalmanagement.pharm_pat_details', id= req["pat_id"])
+
+    return abort(400)
 
 @bp.route('/<string:section>/psearch', methods =['POST'])
 @login_required
@@ -121,12 +150,15 @@ def search(section):
                 return redirect(url_for('hospitalmanagement.diag_pat_details', id = id))
             elif(section=='pharmacyPatSearch'):
                 return redirect(url_for('hospitalmanagement.pharm_pat_details', id = id))
+            elif(section=='billPatSearch'):
+                return redirect(url_for('hospitalmanagement.bill_pat_details', id = id))
             else:
                 return abort(404)
 
     return abort(400)
 
 @bp.route('/viewpatients', methods = ['GET'])
+@login_required
 def view():
     if request.method == 'GET':
         pat = PatientStore.objects(ws_status= 1)
@@ -138,49 +170,85 @@ def view():
     return abort(400)
 
 @bp.route('/update', methods=['POST','GET'])
+@login_required
 def update():
+    patient=None
     if request.method == 'POST':
         if request.form['submit_button'] == 'Get_button':
 
             patient = PatientStore.objects(ws_pat_id= request.form['patient_id']).first()
             if patient==None:
-                flash('Invalid Patient ID', "error")
-                return render_template('update.html')
+                flash('Invalid Patient ID', "danger")
+                return render_template('update.html', patient= patient)
             else:
                 patient = PatientStore.objects(ws_pat_id= request.form['patient_id']).first()
                 return render_template('update.html', patient = patient)
             
         elif request.form['submit_button'] == 'Update_button':
-            PatientStore.objects(ws_pat_id= request.form['patient_id']).update(patient_name = request.args.get('pat_name'),age = request.args.get('pat_age'),doj = request.args.get('pat_doa'),rtype = request.args.get('bed-type'),adrs = request.args.get('pat_address'),state = request.args.get('pat_state'),city = request.args.get('pat_city'))
-            flash('Updated Successfully', "success")
-            return render_template('update.html')
+
+            try:
+
+                PatientStore.objects(ws_pat_id= request.form['patient_id']).update(ws_pat_name = request.form['pat_name'],
+                ws_age = request.form['pat_age'],ws_doj = datetime.strptime(request.form['pat_doa'], '%Y-%m-%dT%H:%M'),
+                ws_rtype = request.form['bed_type'],ws_adrs = request.form['pat_address'],ws_state = request.form['pat_state'],
+                ws_city = request.form['pat_city'])
+                if PatientStore.objects(ws_pat_id= request.form['patient_id']).count()==1:
+                    flash('Updated Successfully', "success")
+                else:
+                    raise Exception('Wrong Patient ID')
+            except:
+                flash('Update failed. Do not change Patient ID and ensure all the data is filled.', 'danger')
+            return render_template('update.html', patient= patient)
 
     else:
-        return render_template("update.html")
+        
+        return render_template("update.html", patient= patient)
 
 @bp.route('/delete', methods=['POST','GET'])
+@login_required
 def delete():
+
+    patient= None
     if request.method == 'POST':
+        
         if request.form['submit_button'] == 'Get_button':
 
             patient = PatientStore.objects(ws_pat_id= request.form['patient_id']).first()
             if patient==None:
-                flash('Invalid Patient ID', "error")
+                flash('Invalid Patient ID', "danger")
                 return render_template('delete.html')
             else:
                 patient = PatientStore.objects(ws_pat_id= request.form['patient_id']).first()
                 return render_template('delete.html', patient = patient)
 
         elif request.form['submit_button'] == 'Delete_button':
-            PatientStore.objects(ws_pat_id= request.form['patient_id']).delete()
-            flash('Deleted Successfully', "success")
-            return render_template('delete.html')
+
+            try:
+                if PatientStore.objects(ws_pat_id= request.form['patient_id']).count()==1:
+                    PatientStore.objects(ws_pat_id= request.form['patient_id']).delete()
+                    flash('Deleted Successfully', "success")
+                
+                else:
+                    raise Exception('Delete Exception')
+            except:
+                flash('Delete failed. Do not change Patient ID', 'danger')
+
+            return render_template('delete.html', patient = patient)
         
     else:
-        return render_template("delete.html")
+        return render_template("delete.html", patient=patient)
+
+@bp.route('/bill', methods=['GET', 'POST'])
+@login_required
+def bill_pat_details():
+    pat= None
+    meds= None
+    diag= None
+
+    return render_template('bill.html', pat= pat, meds= meds, diag= diag)
     
 @bp.route('/logout')
 @login_required
 def logout():
     session.clear()
-    return redirect(url_for(home))
+    return redirect(url_for('hospitalmanagement.home'))
